画像の連続取得&画像処理実装

+++ tracker/tracking/components/Rectangle.py
@@ -1,7 +1,7 @@
 from components.Point2i import Point2i
 
 class Rectangle():
-    def __init__(self, upperLeftPoint:Point2i, width:float, height:float):
+    def __init__(self, upperLeftPoint:Point2i, width:int, height:int):
         self.UpperLeftPoint:Point2i = upperLeftPoint
-        self.Width:float = width
-        self.Height:float = height
+        self.Width:int = width
+        self.Height:int = height
